refactor: extract projection list componet
